Add files via upload
An insurance company in the US is reviewing its insurance claims/charges and is trying to do a cause-and-effect analysis for future business decisions. It has collected data for its customers who have made claims in recent times. The data points collected are age, gender, BMI, number of children/dependents, smoking habit, region they belong to, and charges/bills claimed under the insurance. This analysis would have a bearing on what premium the company should charge a customer availing an insurance policy.