Add files via upload
Terro’s Real Estate is an agency that estimates the pricing of houses in a certain locality. The pricing is concluded based on different features/factors of a property. This also helps them identify the business value of a property. To do this activity the company employs an “Auditor”, who studies various geographic features of a property like pollution level (NOX), crime rate, education facilities (pupil to teacher ratio), connectivity (distance from highway), etc. This helps in determining the price of a property.